chore: 설정 파일 업로드 및 여러 불필요한 파일 제거와 리팩토링 작업 수행
