+++ .gitignore
@@ -35,4 +35,3 @@ out/
 
 ### VS Code ###
 .vscode/
-/src/main/resources/application.yml
