+++ src/main/java/com/bob/mate/domain/post/entity/Post.java
@@ -1,0 +1,35 @@
+package com.bob.mate.domain.post.entity;
+
+import com.bob.mate.global.embedded.TimeColumns;
+import lombok.*;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+import static javax.persistence.CascadeType.*;
+import static javax.persistence.GenerationType.IDENTITY;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "post")
+public class Post {
+
+    @Id
+    @Column(name = "post_id")
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    @Column(length = 50, nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String content;
+
+    @OneToMany(mappedBy = "post", cascade = ALL)
+    private List<Tag> tags = new ArrayList<>();
+
+    @Embedded
+    private TimeColumns timeColumns;
+}
