+++ src/main/java/com/bob/mate/domain/post/entity/Tag.java
@@ -1,0 +1,30 @@
+package com.bob.mate.domain.post.entity;
+
+import com.bob.mate.global.embedded.TimeColumns;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "tag")
+public class Tag {
+
+    @Id
+    @Column(name = "tag_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "tag_name")
+    private String name;
+
+    @ManyToOne(fetch = FetchType.LAZY,cascade = CascadeType.ALL)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @Embedded
+    private TimeColumns timeColumns;
+}
