--- src/main/java/com/bob/mate/domain/post/repository/PostRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.bob.mate.domain.post.repository;
-
-import com.bob.mate.domain.post.dto.PostQueryDto;
-
-import java.util.List;
-
-public interface PostRepositoryCustom {
-
-    List<PostQueryDto> findAllPost();
-}
