+++ src/main/java/com/bob/mate/domain/post/repository/TagRepository.java
@@ -1,10 +1,7 @@
 package com.bob.mate.domain.post.repository;
 
-import com.bob.mate.domain.post.domain.Tag;
+import com.bob.mate.domain.post.entity.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface TagRepository extends JpaRepository<Tag, Long> {
-    Optional<Tag> findByName(String name);
 }
