+++ src/main/java/com/bob/mate/domain/user/controller/UserController.java
@@ -1,6 +1,5 @@
 package com.bob.mate.domain.user.controller;
 
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -8,14 +7,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-
 @RequestMapping("/user")
 @RequiredArgsConstructor
 @Slf4j
 @Controller
 public class UserController {
-
-
 
     @GetMapping("/login")
     public String userLogin(Model model) {
