+++ src/main/java/com/bob/mate/domain/user/entity/Role.java
@@ -1,6 +1,5 @@
 package com.bob.mate.domain.user.entity;
 
 public enum Role {
-
     USER, ADMIN
 }
