+++ src/main/java/com/bob/mate/domain/user/entity/User.java
@@ -1,27 +1,24 @@
 package com.bob.mate.domain.user.entity;
 
-import com.bob.mate.global.audit.AuditListener;
-import com.bob.mate.global.audit.Auditable;
-import com.bob.mate.global.audit.TimeEntity;
-import lombok.Builder;
+import com.bob.mate.global.embedded.TimeColumns;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
-import static javax.persistence.EnumType.*;
-import static javax.persistence.GenerationType.*;
-import static lombok.AccessLevel.*;
+import static javax.persistence.EnumType.STRING;
+import static javax.persistence.GenerationType.IDENTITY;
+import static lombok.AccessLevel.PROTECTED;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = PROTECTED)
-@EntityListeners(AuditListener.class)
-public class User implements Auditable {
+@Table(name = "member")
+public class User {
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
-    @Column(name = "user_id")
+    @Column(name = "member_id")
     private Long id;
 
     @Column(nullable = false, unique = true)
@@ -33,24 +30,6 @@ public class User implements Auditable {
     private Role role;
 
     @Embedded
-    private TimeEntity timeEntity;
-
-
-    @Builder
-    public User(String username, String password, Role role, TimeEntity timeEntity) {
-        this.username = username;
-        this.password = password;
-        this.role = role;
-        this.timeEntity = timeEntity;
-    }
-
-
-    /**
-     * AuditListener
-     */
-    public void setTimeEntity(TimeEntity timeEntity) {
-        this.timeEntity = timeEntity;
-    }
-
+    private TimeColumns timeColumns;
 }
