--- src/main/java/com/bob/mate/global/audit/AuditListener.java
@@ -1,34 +1,0 @@
-package com.bob.mate.global.audit;
-
-import org.springframework.context.annotation.Configuration;
-
-import javax.persistence.PrePersist;
-import javax.persistence.PreUpdate;
-import java.time.LocalDateTime;
-
-
-@Configuration
-public class AuditListener {
-
-    @PrePersist
-    public void setCreateDate(Auditable auditable) {
-        TimeEntity timeEntity = auditable.getTimeEntity();
-
-        if (timeEntity == null) {
-             timeEntity = new TimeEntity();
-            auditable.setTimeEntity(timeEntity);
-        }
-        timeEntity.setCreatedDate(LocalDateTime.now());
-
-
-    }
-
-
-    @PreUpdate
-    public void setUpdateDate(Auditable auditable){
-        TimeEntity timeEntity = auditable.getTimeEntity();
-        timeEntity.setUpdatedDate(LocalDateTime.now());
-
-    }
-
-}
