--- src/main/java/com/bob/mate/global/audit/Auditable.java
@@ -1,8 +1,0 @@
-package com.bob.mate.global.audit;
-
-public interface Auditable {
-
-    TimeEntity getTimeEntity();
-
-    void setTimeEntity(TimeEntity time);
-}
