--- src/main/java/com/bob/mate/global/audit/TimeEntity.java
@@ -1,22 +1,0 @@
-package com.bob.mate.global.audit;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
-import java.time.LocalDateTime;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@Embeddable
-public class TimeEntity {
-
-    @Column(updatable = false)
-    private LocalDateTime createdDate;
-
-    private LocalDateTime updatedDate;
-
-}
