+++ src/main/java/com/bob/mate/global/config/PrincipalDetailsService.java
@@ -1,10 +1,7 @@
 package com.bob.mate.global.config;
 
-import com.bob.mate.domain.user.entity.Role;
-import com.bob.mate.domain.user.entity.User;
 import com.bob.mate.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -14,24 +11,12 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-@Slf4j
 public class PrincipalDetailsService implements UserDetailsService {
 
     private final UserService userService;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-
-        log.info("userId = {}", username);
-        User userEntity = userService.findByUsername(username);
-
-
-        log.info("userEntity  = {}", userEntity);
-
-        if (userEntity != null) {
-            return new PrincipalDetails(userEntity);
-        }
         return null;
-
     }
 }
