+++ src/main/resources/application.yml
@@ -1,0 +1,24 @@
+spring:
+  datasource:
+    url: ${{secrets.POSTGRES_DATABASE}}
+    username: ${{secrets.POSTGRES_USERNAME}}
+    password: ${{secrets.POSTGRES_PASSWORD}}
+    hikari:
+      maximum-pool-size: 4
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.PostgreSQLDialect
+        format_sql: true
+        use_sql_comments: true
+
+logging:
+  level:
+    org:
+      hibernate:
+        type:
+          descriptor:
+            sql: trace
