--- src/main/resources/static/css/chat.css
@@ -1,19 +1,0 @@
-*{ margin: 0; padding: 0; }
-
-.chat_wrap .header { font-size: 14px; padding: 15px 0; background: #F18C7E; color: white; text-align: center;  }
-
-.chat_wrap .chat { padding-bottom: 80px; }
-.chat_wrap .chat ul { width: 100%; list-style: none; }
-.chat_wrap .chat ul li { width: 100%; }
-.chat_wrap .chat ul li.left { text-align: left; }
-.chat_wrap .chat ul li.right { text-align: right; }
-
-.chat_wrap .chat ul li > div { font-size: 13px;  }
-.chat_wrap .chat ul li > div.sender { margin: 10px 20px 0 20px; font-weight: bold; }
-.chat_wrap .chat ul li > div.message { display: inline-block; word-break:break-all; margin: 5px 20px; max-width: 75%; border: 1px solid #888; padding: 10px; border-radius: 5px; background-color: #FCFCFC; color: #555; text-align: left; }
-
-.chat_wrap .input-div { position: fixed; bottom: 0; width: 100%; background-color: #FFF; text-align: center; border-top: 1px solid #F18C7E; }
-.chat_wrap .input-div > textarea { width: 100%; height: 80px; border: none; padding: 10px; }
-
-.format { display: none; }
-
