--- src/main/resources/templates/chat.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-
-<head>
-  <link th:href="@{/css/chat.css}" rel="stylesheet">
-
-  <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css"
-        integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-</head>
-<body>
-
-<div class="chat_wrap">
-  <div class="header">
-    CHAT
-  </div>
-  <div class="chat">
-    <ul>
-      <!-- 동적 생성 -->
-    </ul>
-  </div>
-  <div class="input-div">
-    <textarea placeholder="Press Enter for send message."></textarea>
-  </div>
-
-  <!-- format -->
-
-  <div class="chat format">
-    <ul>
-      <li>
-        <div class="sender">
-          <span></span>
-        </div>
-        <div class="message">
-          <span></span>
-        </div>
-      </li>
-    </ul>
-  </div>
-</div>
-
-<script th:src="@{/js/chat.js}"></script>
-</body>
-
-
-</html>
-
-
