--- src/main/resources/templates/user/login.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css"
-          integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
-    <meta charset="UTF-8">
-    <title>Login</title>
-</head>
-
-<body class="text-center">
-
-<main class="form-signin">
-<form class="form-signin" th:action="@{/user/login}" method="post">
-    <a th:href="@{/}"><img class="mb-4" src="https://getbootstrap.com/docs/5.1/assets/brand/bootstrap-logo.svg" alt="" width="72" height="72"></a>
-    <h1 class="h3 mb-3 fw-normal">Please sign in</h1>
-
-    <label for="username" class="sr-only">Username</label>
-    <input type="text" class="form-control" id="username" name="username" placeholder="Id">
-
-    <label for="password" class="sr-only">Password</label>
-    <input type="password" class="form-control" id="password" name="password" placeholder="Password">
-
-    <button class="w-100 btn btn-lg btn-primary" type="submit">Sign in</button>
-
-</form>
-</main>
-</body>>
-</html>
